nzemn_v1/nzemn1/requirements.txt
django>=2.1
psycopg2
django-guardian
djangorestframework
djangorestframework-gis
django-filter
pyyaml
uritemplate
